+++ Sources/AjarCore/Edit/MediaReferenceEditCommands.swift
@@ -3,6 +3,25 @@
 import Foundation
 
 extension EditReducer {
+    static func addMediaReferences(
+        _ additions: [MediaRef],
+        to project: Project
+    ) throws -> Project {
+        var seenIDs = Set(project.mediaPool.map(\.id))
+        for addition in additions where !seenIDs.insert(addition.id).inserted {
+            throw EditReducerError.duplicateImportedMediaReferenceID(addition.id)
+        }
+
+        return Project(
+            schemaVersion: project.schemaVersion,
+            schemaMinor: project.schemaMinor,
+            settings: project.settings,
+            mediaPool: project.mediaPool + additions,
+            sequences: project.sequences,
+            looks: project.looks
+        )
+    }
+
     static func updateMediaReferences(
         _ replacements: [MediaRef],
         in project: Project